+++ src/App.css
@@ -2,15 +2,15 @@ body {
   background-color: #d9f7ff;
 }
 
+/**********************
+Board layout and design
+**********************/
+
 #title {
   font-family: 'Share', cursive;
   font-size: 1.5em;
   letter-spacing: 3px;
-}
-
-.container {
-  margin-top: 20px;
-  align-items: center;
+  margin: 20px 0;
 }
 
 .row {
@@ -18,20 +18,24 @@ body {
 }
 
 .square {
-  height: 60px;
   background-color: white;
   text-align: center;
-  padding: 5px;
-  font-size: 2em;
   border: 1px solid black;
+  font-size: 2em;
+  padding: 5px;
+  height: 60px;
 }
 
-.status {
-  font-size: 1.5em;
-  width: 100%;
+.info {
   text-align: center;
+  font-size: 1.5em;
   margin-top: 10px;
+  width: 100%;
 }
+
+/************
+media queries
+************/
 
 @media only screen and (min-width: 767.5px) {
   .square {
